backend/requirements.txt
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe
python-dotenv==0.17.1
Werkzeug==2.2.3
dataclasses-json==0.5.7
distro==1.7.0
dnspython==2.3.0
Flask-Cors==3.0.10
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.6.0
langchain==0.0.194
langchainplus-sdk==0.0.6
oauthlib==3.2.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pexpect==4.8.0
pinecone-client==2.2.2
protobuf==3.20.0
python-dotenv